+++ ffi/src/sspi/mod.rs
@@ -3,12 +3,12 @@ mod macros;
 
 pub mod common;
 pub mod credentials_attributes;
+pub mod scard_cert;
 pub mod sec_buffer;
 pub mod sec_handle;
 pub mod sec_pkg_info;
 pub mod sec_winnt_auth_identity;
 pub mod security_tables;
-#[cfg(feature = "scard")]
-pub mod smartcard;
 pub mod sspi_data_types;
 pub mod utils;
+pub mod win_scard_cert;
